Expiring metadata state: check item in map before deleting
Summary: For all of our maps, we should check that the item in the map is the correct item before deleting it. otherwise, if a pod starts up with the same name as one that expired less than 24 hours ago, it would be deleted once that first pod expires.

Test Plan: n/a

Reviewers: vihang, nserrino, zasgar

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10027

GitOrigin-RevId: 19bec3ee39f7016174a7c0ec588d51abee213927