PC-1221: Clean up signup logic
Summary:
Signup will be changed with the addition of IdentityProviderOrgName logic so I'm cleaning up this function because I found it hard to read. A big issue we had is that we instantiated variables at the top of the function and had them defined at some if statement. I had a hard time telling who actually did the work. Now we always return when we are inside an if statement by calling `signupUser()` with new info - no more magic.

Changed one piece of logic: go back to using the domainName of a user's email instead of their IdentityProviderOrgName. In a follow up diff, will check the IdentityProviderOrgName of the user for org membership. Note: this only happens if the IPON exists for the user, otherwise they're put into a self-org so it's slightly different than the original org logic.

Test Plan: Updated tests and skaffold deployed.

Reviewers: vihang

Reviewed By: vihang

JIRA Issues: PC-1221

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10053

GitOrigin-RevId: 16d4b41b0b964d3fbdb051b8565f134511f8a9de