Replace stale resource memory usage script with service resource usage script
Summary: The previous version of this script gives service memory usage as an output table over time, which isn't that useful because a table format is better suited to per-service high-level stats. @htroisi recently put out a YouTube video showing how to use Pixie to help analyze whether or not your services are over or underprovisioned. We want to provide a link to the script, which shows resource utilization and request throughput/latency by service, in the video description. Since it's a more interesting script but related to the stale one, just replacing the stale one here.

Test Plan: ran the script in the UI

Reviewers: oazizi, zasgar, philkuz

Reviewed By: oazizi, zasgar

Subscribers: htroisi

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10069

GitOrigin-RevId: 1b849c9dbed397a480b5eef84df3887a33fd2982