Fix bug in planner that was seen with recent changes to px/cluster.
Summary:
This diff fixes a bug in IR::IndependentGraphs, where not all nodes are included in the independent graphs it outputs. This leads to strange downstream behaviour such as a GRPCSource on a PEM.
This diff does not fix the non-determinism of the planner, but we should do that at some point.

I also fixed IR::DebugString to be const, since that bothered me while I was debugging.

Test Plan: Added ir_test.cc with tests for IndependentGraph.

Reviewers: nserrino, philkuz

Reviewed By: philkuz

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10089

GitOrigin-RevId: ee7b4eeb5c533f31d0d553da32e9dc4d7134c8ad