UI: Remove some cruft and update naming in deep linking logic
Summary:
Some of the old logic was stale and convoluded. It also made it harder to add a deep link to px/ip for IP addresses without giving px/ip a vanity URL (which doesn't make sense for this particular case).

What we really want is two ways to create a deep link: Deep link from a semantic type+value pair, and Deep link to a particular script.

In this diff, I clean up some of the unused exported fields, and update the naming on some of the functions. In the next diff, I update the implementation and interface of the deep linking logic a bit more (keeping parity with current functionality). In the next diff, I add a deep link to px/ip (which isn't possible to do without a vanity URL in the current code setup).

We might actually want to just remove vanity URLs together, but that is a separate decision that we can make later (either way, cleaning up some of this logic makes the code more extensible for future deep links).

Test Plan: unit tests, run the UI

Reviewers: nlanam, michelle, vihang

Reviewed By: nlanam

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10092

GitOrigin-RevId: e6a76831e6c9a1d6d464240fd3aaa1f24ff8b462