[PC-1246] Update React from v16 to v17
Summary: This is a maintenance release for React, with no breaking changes to the API. It's all internal refactoring to make future upgrades easier, so it's a drop-in upgrade for us. Doing this because Material v5 requires it.

Test Plan: Build. Smoke test. Shouldn't notice anything different.

Reviewers: michelle, vihang, #third_party_approvers

Reviewed By: vihang, #third_party_approvers

JIRA Issues: PC-1246

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10129

GitOrigin-RevId: 731605ef42e5ed8c95de4a230493e2073949c9f2