Cleanup userID variable and fix-up login tests
Summary: We pass around this userID variable in login.go, but this has been available in UserInfo struct for quite a while now as the same value. To reduce confusion, we now just use the userInfo.AuthProviderID value in place of the userID. Evidence this was confusing shows up in the tests as well -we would use the PL User ID and authProviderID interchangably. This fixes that issue in the test.

Test Plan: Tests still pass, can still login with skaffold

Reviewers: vihang, michelle

Reviewed By: vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10154

GitOrigin-RevId: 05e2e3ad7e5514b4a3271efa47cde8f6c9ca4f60