Allow users to login who lack an org
Summary:
Users who would tried to call GetAugmentedToken but lacked an org would receive an error. This call is made as part of the GRPC middleware in the api_server. We will need to call GetUser/ GetOrg() from the UI in these cases so probably just allow users to login without an org.

This diff is part of a series to enable user/password signup flows. We will need to allow users to exist in the database without an OrgID and make sure that the data is accessible to users.

Risk - we are creating a new opening for users to log in without an org ID. We might provide more access than we intend. However, most features are org-gated anyways - meaning we restrict info so that only the org members can view clusters, org members, so those risks will be limited.

Test Plan: Updated the test. No chance to do an integration because we don't support orgless users yet.

Reviewers: vihang, michelle, nserrino

Reviewed By: vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10184

GitOrigin-RevId: b199c233d18d840e224b567a9c21cccaa971a0dc