PC-1248 Rm vizier_cluster first_seen_at col
Summary:
this column was preemptively created during table creation, but is unused. it would actually contain the same information as the `created_at` col that also exists in the table.
since we're not using it, we might as well remove it for cleanliness

Test Plan: stuff still builds. there is no use of this anywhere in the code

Reviewers: nserrino, vihang, zasgar

Reviewed By: nserrino, vihang

JIRA Issues: PC-1248

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10199

GitOrigin-RevId: 38a9608c1a053b6ca775b73dbd2f5749b54f5c0f