[PC-1273] Create org setup page (no-op)
Summary:
Creates a `/setup` route, that intercepts other routes when the user logs in but still has to set up their organization.

This diff leaves stubs for the following behavior. There is a TODO at each site to explain what to change.
- Deciding if setup is complete: currently assumes it is.
- Validation of org name: just checks that it's at least three characters long.
- Actually creating the org: pretends that it did, then redirects a moment later.

Seeing the page:
{F173137}

Invalid input:
{F173138}

Valid input:
{F173143}

Test Plan:
To see the setup page, go to `containers/App/live.tsx` and force `setupComplete` to be false. When the page refreshes, all non-auth-related requests should force their way to the new `/setup` route, with a valid `redirect_uri`. The button labeled `Create` becomes enabled when the org name is typed in and has at least two characters. Clicking the button shows a message, then tries to redirect to the place you started.
When the setup IS marked as complete, attempting to visit the setup page will immediately follow the redirect instead, and setup will not intercept attempts to go to other routes.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-1273

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10200

GitOrigin-RevId: 477932fe1a62894cd547786dc219d6c22d3cb197