Explicit permission for orgs in org_grpc
Summary:
Among the grpc requests we make, org_grpc is the only one that allows users to pass in an org_id. We dont' want to allow users to get info about other orgs via the interface, but the interface implies that they can. The implementations currently block any such requests. That eventually might change so this is an important step to prevent users from a future potential leak.

This diff just adds a comparison to make sure the claimsID matches the requested ID.

Test Plan: Added test set that check whether the functions will be blocked when the user make requests for other orgs and allowed when the user requests for their own org.

Reviewers: michelle, vihang

Reviewed By: michelle

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10202

GitOrigin-RevId: 59d6e4e6c47b63eab9d8cf442bfaf9d89b38f335