Fixes to login code to support org creation
Summary:
Login heavily relies on PLOrgID for a lot of things. However with the
new email password based users, org creation happens later. So switch to fetching
the user using the AuthProviderID and then looking at the orgID from our database
instead.
Ensure that this orgID is also included in the JWT from the login response.

Test Plan: Existing tests

Reviewers: michelle, philkuz

Reviewed By: michelle

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10218

GitOrigin-RevId: 5f394d5afb206491ed337d74645e7ea84be0ae8f