Call CreateOrg() mutation when CreateOrg called.
Summary:
Part of a series of diffs to make CreateOrg work. This one wires up the frontend
with the gql calls necessary. Also makes a post request to the backend after the call succeeds to refresh the claims on the auth token so that the user will be properly authorized.

Test Plan: Tested against skaffold backend and able to login.

Reviewers: vihang, michelle

Reviewed By: vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10220

GitOrigin-RevId: ba9bc293c17d05467b91d7fc9eac4fcf9a8a7851