[PC-1283] Simplify canvas resize logic
Summary: This was an incomplete refactor that had a missing null check, which a user ran into in production. I tried a further simplification (the TODO that this diff removes), but that caused different problems.

Test Plan: Run a script. Switch scripts. Resize the window both horizontally and vertically. Widgets should still resize as needed.

Reviewers: michelle, vihang

Reviewed By: vihang

JIRA Issues: PC-1283

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10248

GitOrigin-RevId: e3e924c5e408e39c7aa32fd5ce102979171cb696