PC-1281: Remove duplicate script execution started analytics event from `px run`
Summary: This event is a duplicate, which is causing the number of script execution events to be double what it should be. It also exists in script.go, so we are deleting this event and relying on that one.

Test Plan: Ran bazel px CLI and printed out the events that are sent.

Reviewers: zasgar, michelle, vihang, philkuz

Reviewed By: vihang

JIRA Issues: PC-1281

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10265

GitOrigin-RevId: 3785a218267a88b90c7e552d8ce1f08b9fee53ed