Dynamic Tracer: Better support for suffix symbol matches
Summary:
This fixes a test that fails locally on my machine, presumably because an upgrade caused different versions of shared libraries to be installed.

The new libraries exposed a flaw in how we search for suffix symbol matches.

Test Plan: The existing tests. Tested manually on my local machine and on Jenkins.

Reviewers: #stirling, yzhao

Reviewed By: #stirling, yzhao

Subscribers: yzhao

Signed-off-by: Omid Azizi <oazizi@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10273

GitOrigin-RevId: 1e120885f2ee65c5726e78cecc209e9095975341