Remove org_name from login callback
Summary:
Was reading over the code as part of investigating the invite flow and discovered that
we still had a parameter field for org_name in the client calls, but the server doesn't actually
care about the orgName. (see src/api/controllers/auth.go). Removing those calls for simplicity.

Test Plan: yarn dev, try to login to prod instance and find that we don't actually get access.

Reviewers: vihang, michelle, nlanam

Reviewed By: vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10284

GitOrigin-RevId: c475d7bb024df288cb3b69100c7b9d4ddac9ce83