Unify callback URL logic in one file
Summary:
I noticed that we separated out logic to create callback urls from the logic that parses the callback urls. I figured we should unify them in one place to make it clearer that the data is related.

This is purely a move and some interface modifications. Cleanup like removing dead fields will happen in a later diff.

Test Plan: Try all login flows to make sure that the data is identical.

Reviewers: vihang, michelle, nlanam

Reviewed By: vihang

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10296

GitOrigin-RevId: 40916fe629d9cf3f819036a5fb8af8ac74750e12