[PC-1301] Create UI for invitation links
Summary:
{F174453}

- Adds an "Invite Users" button in the Users tab of the Admin view.
- The "Invite Users" button summons a modal with an invitation link.

This does NOT add the following (those will be separate diffs):
- UI to regenerate invitation links
- UI to remove users
- Signup page changes for invitation links

NOTE: This is behind a feature flag named `invite`. Enable that or force it in `invite-user-button.tsx` to test further change.

Test Plan: Open the users tab in Admin. There should be an "Invite Users" button to the right of the tab bar. Clicking it opens a modal. Clicking the "Copy link" button in that modal copies the link. The text field can't be edited, but focusing it highlights its contents.

Reviewers: michelle, vihang

Reviewed By: michelle

Subscribers: philkuz

JIRA Issues: PC-1301

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10322

GitOrigin-RevId: 72e8b0f3b3b319be2801e25c3eabbedcde2baf5a