PC-1311: Support Terminated pod phase in Vizier
Summary: We want to extend the pod phases we track in autocomplete and Vizier in order to include a terminated state. This is so that autocomplete and Vizier queries can report a "Terminated" state for pods whose last known kubernetes phase was "running" but are in fact terminated by the system. (Failed and Succeeded pods will continue to have the same phase as before). In this diff, we add the terminated pod phase to Vizier (both in the metadata service as well as in the metadata UDFs). This is an extension of the Kubernetes pod phases, see PC-1311 and PC-1305 for more context.

Test Plan: Ran Vizier against prod cloud and deleted a pod, unit tests

Reviewers: michelle, philkuz

Reviewed By: philkuz

JIRA Issues: PC-1311

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10339

GitOrigin-RevId: 4ca6f1fdaf3e99f5f8a4dc6adbad5edc24f25744