PC-1322 Update indexer to derive node state from node conditions
Summary:
As stated in D10362, we were incorrectly using the node phase to determine node state. Instead, we need to rely on the node conditions to derive the
current state of the node. This diff updates the indexer so that it reads the nodeConditions to determine a state.

Test Plan: unit test

Reviewers: nserrino, vihang

Reviewed By: vihang

JIRA Issues: PC-1322

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10363

GitOrigin-RevId: 4384ee6981c88c1b4a8f974a9a3c7157af4fd9db