[PC-1335] Fix CreateOrg page not showing up for org-less users
Summary:
CreateOrg page wasn't opening because somewhere along a different path we switched the logic to check if orgData exists
rather than if the orgName exists. This switches it back.

Test Plan: Tested on yarn with staging backend and saw a createOrg page when I expected it.

Reviewers: vihang, michelle, nlanam

Reviewed By: vihang

JIRA Issues: PC-1335

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10404

GitOrigin-RevId: 08780ab4e7b0036ccd4e5d3b36f8cc71d07e4cd3