Add clusterID to cloudconnector annotations
Summary:
the cluster ID is currently saved as a secret in `pl-cluster-secrets`. this makes it annoying for other applications (for example the CLI) to get the cluster ID, when it doesn't actually need to be a sensitive secret.
this change makes the cloudconnector expose the clusterID as an annotation.

Test Plan: deployed via Skaffold and checked that the pod contains the cluster id in the annotations

Reviewers: vihang, zasgar

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10405

GitOrigin-RevId: 34e33edf888a47f79d7b9a89a4c2932c292392ca