[PP-3155] Force vega tooltip inside window bounds.
Summary:
Also forces wrapping on long titles (flamegraphs show generated method signatures, for example), and limits that wrapping to four lines before an ellipsis.

This is done in a global style, as the issue affects all Vega tooltips (there seems to be a bug upstream in position calculation).

Here's a worst-case scenario after the fix. Very long title, Vega decided to put it on the left (and tried to push it off-screen); the CSS overrides forced it back to reality.
{F175429}

Test Plan: Run `px/perf_flamegraph` and find tooltips with ridiculously long titles. Move the mouse around; it may take a while to find one that would try to force the tooltip off-screen.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PP-3155

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10415

GitOrigin-RevId: 1819419bff378183d740343eb42562cf1dcd4909