PC-1308 Deduplicate results in autocomplete
Summary:
We currently don't deduplicate results in autocomplete, so if there is statefulset which has restarted with a different UID, it will show up twice. instead, we only want to show the most recent statefulset.
this diff uses elastic's collapse functionality to deduplicate results, and ranks the resource with the most recent updateVersion at the top.

Test Plan: updated the suggester test to include two resources with the same name. ensure that the resource with the later updateVersion is returned in the results.

Reviewers: nserrino, vihang

Reviewed By: vihang

JIRA Issues: PC-1308

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10428

GitOrigin-RevId: dc3be4f15178d8733996774ba4cea1b1175cf19c