Log VizierQueryError details to console
Summary:
Not a real fix, but this at least gives a savvy user a bit more info than "your vis spec is wrong" when the real problem was that the script refers to a variable that isn't in the vis spec.

Split from another unrelated diff.

Test Plan: Run `px/http_data`. Remove the `source_filter` variable from the vis spec. Run again. You get the same not-so-helpful snackbar, but the real error goes to the browser console.

Reviewers: michelle, vihang

Reviewed By: michelle

Subscribers: philkuz

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10430

GitOrigin-RevId: 889f7a7d1c681085a60b9b9bef7014ef1a3f5ea7