[PC-1343] Remove keyboard shortcut for Command Input
Summary: It was already a no-op, but it was still appearing in command help. This does not remove the component; that still exists albeit as dead code.

Test Plan: Open Live UI. Use `Shift`+`/` to open keyboard shortcut help. Note there is no longer an entry for the Command Input; the old shortcut (`Ctrl/Cmd`+`K`) is no longer consumed by the UI. Instead, the default browser behavior happens (in Chrome and Firefox, it jumps to the URL bar in search mode).

Reviewers: michelle, vihang, philkuz

Reviewed By: philkuz

JIRA Issues: PC-1343

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10432

GitOrigin-RevId: db924388260225a6395a7c80fed604b3dac3c6d3