[PP-3170] Allow setting the total data stored by the table store.
Summary:
This diff allows setting the total amount of data stored across all tables by the table store, as a flag/envvar on PEM.
In the future, we will expose this to the user somehow but that part is not clear yet.
This diff also allows setting how much of the total limit should be allocated to the http_events table. This is also a flag/envvar but there are no current plans to expose this further to the user.
The default values are chosen such that the amount of data for each table is the same as prior to this diff as long as the flag/envvar is not set.

Test Plan: Tried without setting envvar, and then ran `px._DebugTableInfo()` to see that the values are the same as prior to this diff. Changed the envvar and re-ran to make sure it changes the table max sizes.

Reviewers: nserrino

Reviewed By: nserrino

JIRA Issues: PP-3170

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10449

GitOrigin-RevId: c005676d293b6fb26aa8aae62676d34b39b3f99b