PC-1354: Improve error handling/logging of the cloud indexer's Vizier metadata watcher
Summary:
We've noticed that certain clusters fail to send any indexer data to Elastic on Pixie Cloud. This persists even in the case where we restart the indexer in cloud or restart Vizier fon that cluster. We've also seen logs about stan subscription timeouts in the indexer. In addition, there are repeated errors of Viziers trying to reconnect to the indexer, even though the indexer supposedly already started listening to those Viziers.

It seems the following might be happening,
1. Vizier tries to connect to the indexer to send over metadata
2. The metadata subscription fails due to timeout
3. Future attempts by that Vizier to connect to the indexer are blocked, because the indexer believes that the connection with that Vizier is already being handled (even though it's not because it failed to subscribe to stan).

In this diff, the error from the stan subscription should now be propagated up and attributed to a particular Vizier so it's easier to diagnose if that's happening. In addition, if the indexer fails to establish a stan connection for a given Vizier, it will now allow subsequent retries for that Vizier rather than blocking them.

This alone won't fix the full issue but it should hopefully address some of it and give us better insight into the cases of Viziers that seem unable to send metadata to the indexer.

Test Plan: will want to run this in testing.

Reviewers: michelle, philkuz, vihang

Reviewed By: vihang

JIRA Issues: PC-1354

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10440

GitOrigin-RevId: 2e543d161431c75efe3fe2a85330d6368f8b4ab7