Don't show the operatorMissing message if a CRD was ever detected
Summary:
When pixie is being deleted, the CRD might be deleted first and then
the operator would remove the rest of pixie pods during the reconciliation
phase. In such a scenario, the cloud conn reports the last status as degraded
due to the missing CRD. This should prevent that from happening.

Test Plan: skaffold

Reviewers: michelle, nserrino

Reviewed By: michelle

JIRA Issues: PP-3106

Signed-off-by: Vihang Mehta <vihang@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10457

GitOrigin-RevId: 5a17397cefba895aba3a447cdaf663193ef0b6b9