Fix error with invite text field being set to null
Summary:
When you render the invite modal, React doesn't like that the text is null
before the gql is done. Changing to an empty string to avoid that error and we get
the same values.

Got the following errors before:
```
Warning: `value` prop on `input` should not be null. Consider using an empty string to clear the component or `undefined` for uncontrolled components.
```
```
Warning: A component is changing an uncontrolled input to be controlled. This is likely caused by the value changing from undefined to a defined value, which should not happen. Decide between using a controlled or uncontrolled input element for the lifetime of the component. More info: https://reactjs.org/link/controlled-components

```

Test Plan:
yarn dev. before change, opening the modal threw a console error.
After change, opening the modal worked without throwing errors.

Reviewers: nlanam, michelle, vihang

Reviewed By: nlanam

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10463

GitOrigin-RevId: 961265cf9363a4d2468786b8a91a43948b979504