Handle k8s node updates in PEM
Summary: We recently updated our metadata service to send node updates via NATS. We only use these updates in our cloud for autocomplete, but they still get sent to the PEMs. Because the PEMs don't handle these messages, they unnecessarily log per node update. This diff updates the PEM to handle the Node update, but currently doesn't do anything with it yet. We can update that when we have plans to do something with the node updates.

Test Plan: Deployed PEM on skaffold, with logs in HandleNodeUpdate to ensure that function is getting called

Reviewers: nserrino, jamesbartlett, vihang

Reviewed By: vihang

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10472

GitOrigin-RevId: 4092d049a9f9e1f1e1ab0826a9638890df4031ec