Fix flaky table_test.
Summary: The table test was failing every so often because in the threaded test the reader would get ahead of the data and then fail. Instead now the reader will wait for the writer to write more data instead of failing. I also added some comments to make the logic of the test clearer, and added a RAII wrapper around absl::Notification to make sure that the test fails immediately if the writer fails for some reason.

Test Plan: Ran the test 1000 times without a failure. Also checked that if the writer fails for some reason the reader doesn't hang.

Reviewers: zasgar

Reviewed By: zasgar

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10482

GitOrigin-RevId: 8d51cad6d77ca90066cb30a777fc74902817a7a5