Update vzconn manager NATS channel buffer sizes
Summary: We are currently buffering our NATS messages that we read from the channel. As we are currently receiving some slow NATS consumer errors, we would like to increase the buffer size to take som eof the load off of NATS. In a subsequent diff, we will provide a more holistic fix for the slow consumer issue.

Test Plan: n/a

Reviewers: michelle, vihang, zasgar

Reviewed By: michelle

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10491

GitOrigin-RevId: 346cd7d00460233f7380f6c69e9c7fe439e08cc6