VizierWatcher should call onVizier handlers as goroutines, not blocking calls
Summary: VizierWatcher is responsible for watching for Viziers and calling onVizier handlers that are specified by the consuemr of the VizierWatcher. VizierWatcher reads from a NATS channel and currently calls these handlers in a blocking way. However, the handlers themselves might be doing non-trivial work, and we shouldn't block reading from the NATS subscription on those functions. This diff ensures that we wrap onVizier calls in a go routine so that this work does not block the reading from the NATS channel.

Test Plan: n/a

Reviewers: zasgar, vihang, michelle

Reviewed By: vihang

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10497

GitOrigin-RevId: a7f1d973840ad4f9a9fca943b62b420bf06bb199