HTTP: Custom Chunked Transfer Decoder
Summary:
This new custom chunked transfer decoder is a replacement for the existing one by pico.

This version does not mutate the input buffer on a parse failure, which is important because
we may have incomplete data at the time of the parse.

By doing so, we can save a copy of the data buffer which was otherwise required when using pico's parser.

Test Plan:
Numerous test cases added.
Still need to evaluate the performance and memory characteristics of this version.

Reviewers: #stirling, jamesbartlett

Reviewed By: jamesbartlett

Subscribers: yzhao, rcheng, jamesbartlett

Signed-off-by: Omid Azizi <oazizi@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10524

GitOrigin-RevId: a68376d77a912eaa3560c14a9b9ce0c77a6299ca