Fix flakiness in pgsql_trace_bpf_test
Summary:
One of the postgres clients (the go sqlx client) sends the Startup message keys in a different order every time, which results in different behavior in the test.

The variability caused flakiness and also exposed an issue where our BPF inference is not general enough.

This fixes the inference code and detects some missing messsage types for parsing the frames.

Test Plan: Existing tests

Reviewers: #stirling, jps, yzhao

Reviewed By: #stirling, jps, yzhao

Signed-off-by: Omid Azizi <oazizi@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10526

GitOrigin-RevId: fc6c075d4b11f15fe1bb2fea1b402459ed7613b4