PP-3210: Remove a registration check in StartStream that does nothing
Summary:
This registration check is always set to false before we run StartStream, which checks the value in order to run a registration. That means the codepath is always taken on all calls.

I think this only adds cognitive load for the reader and doesn't serve an actual purpose.

Test Plan:
Tested on skaffold. Was able to send data over the Bridge stream to a cloud. Was able
to reconnect during a simulated cloud outage.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PP-3210

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10612

GitOrigin-RevId: ed24df5864ef017a7b2c2340b6b8a154f2eca90f