PC-1397: Bump elastic index again
Summary: We are seeing stale data in our relatively new elastic index, but it's possible that it was put there by some old STAN messages. Let's bump the index again here so we can see whether or not the issue re-occurs (and either way the stale data is incorrect).

Test Plan: n/a

Reviewers: michelle

Reviewed By: michelle

JIRA Issues: PC-1397

Signed-off-by: Natalie Serrino <nserrino@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10566

GitOrigin-RevId: a876f74af321256f0d01e7cba8c7e791b91161c4