PP-3205 Add pod securityContext to vizier spec
Summary:
We have some users who are running with more restrictive pod security policies which require them to specify a securityContext on their pods. this change updates the vizier spec to include the securityContext.
Following changes will use the security context in the vizier spec to determine how the YAML should be modified

Test Plan: none yet, this diff just adds the securityContext to the types but doesn't use it anywhere

Reviewers: vihang

Reviewed By: vihang

JIRA Issues: PP-3205

Signed-off-by: Michelle Nguyen <michellenguyen@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10571

GitOrigin-RevId: 2ee48299c7b32c3395bfd7f5e9b0ce59059d0517