Enable easier local development of external dependencies.
Summary:
There was a lot of pain making changes to bcc, because bazel would rebuild all of bcc and all of bpftrace everytime you made a change.
This diff allows people to locally develop external dependencies and build their artifacts themselves. This way people can take advantage of make/cmake incremental builds while developing locally.
By default, this diff doesn't change the build process at all. But if lines in `repository_locations.bzl` are uncommented, then people can use the new local behaviour for bcc and bpftrace (also if desired other repos, but they aren't setup in this diff).

To achieve this, I add a bazel rule called `local_cc` that mimicks a foreign_cc rule, but instead of running make or cmake, it just copies the artifacts/includes from the specified directories.

This does mean people have to setup the build environment for the external dependencies themselves (when running with the new local repositories, obviously default builds don't have this requirement), but I think people will be happy to do that in exchange for not waiting 3-5 minutes for bcc and bpftrace to rebuild.

Test Plan: Skaffolded with local changes to bcc and bpftrace, and saw that the changes were in effect on the cluster. Skaffolded with the local_repositories commented out, and saw a normal build without the changes to bcc, as expected.

Reviewers: #stirling, vihang, zasgar, #third_party_approvers, yzhao

Reviewed By: #stirling, zasgar, #third_party_approvers, yzhao

Subscribers: yzhao

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10583

GitOrigin-RevId: 6b033ef752b879511de258b51cae34a4fa546a26