DataStream: shrink data stream to retention capacity.
Summary:
After each iteration, the capacity of the data stream buffer's std::string buffer should be reduced to its size.
For example, suppose we pull in, from bpf buffers, 128MB of data, then during `ProcessBytesToFrames` all of this data is successfully processed. The DataStreamBuffer then gets resized to 0, but the capacity of its internal std::string buffer will remain at >128MB.

Test Plan: Tested that the memory is actually freed to the allocator, with the new changes. Tested in skaffold, and in benchmark that is yet to be released.

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

Signed-off-by: James Bartlett <jamesbartlett@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10636

GitOrigin-RevId: 109e1966f85f1c682b083c1d816736fbf1b7544b