PP-3210: [Cleanup] Remove error channel from the vizier-cloud-connector bridge
Summary:
The error channel (`errCh`) in the bridge of vizier-cloud-connector creates
a complex interdependency between the go-routines. I removed this because
it creates a few edge nasty cases on a downstream improvement I'm making and I don't think it
actually adds value.

In the present, the only write occurs when GRPC stream writer errors out. When the error is read from
the channel in HandleNATSBridging, the routine logs and dies. The log can be moved to the source of the error
and the control of the goroutine is redundant. When an error is sent over a GRPC stream, the client connection in gRPC cancels
the context, which sends a message to `stream.Context().Done()`. We listen to `stream.Context().Done()` in all
our goroutines that also listened to `errCh` and die when something arrives. In all cases probed, `stream.Context().Done()` would
recv a message before `errCh` anyways.

Test Plan:
Tested on skaffold by simulating a few errors over the GRPCStream (kill connection, TCP
timeout). The routines died and eventually the system recovers on all situations.

Reviewers: michelle, nserrino, vihang, zasgar

Reviewed By: zasgar

JIRA Issues: PP-3210

Signed-off-by: Phillip Kuznetsov <pkuznetsov@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10626

GitOrigin-RevId: f9b345acc2f97ac1fc0e7f54302bbc850b83fcd0