[PC-1415]: Fixup cache on org creation
Summary:
When creating an org, we weren't reconciling Apollo's cache.
This caused the setup page to loop back on itself when the user created their org.
Logging out and back in (which purges Apollo's cache as a side effect) was the only way to fix it.

This diff just does that reset preemptively.
A more correct fix would be to only `cache.evict` the specific items that need to be refetched, but this fix is good enough as a band-aid for now.

Test Plan: Create a new email/password account. Verify it. Create an org in the UI; the setup page should not return after this and you should have an org that you can interact with.

Reviewers: michelle, vihang

Reviewed By: michelle

JIRA Issues: PC-1415

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10652

GitOrigin-RevId: 7937a67db92b14b1969aa27f0e5dbf34d114c9d1