Memory optimization: Reduce parsed message retention time
Summary:
By reducing the time a message can be stored, we can reduce memory usage.

The new value of 1 minute is still plenty of time for a parsed message to be pushed out as a record.

If it's sticking around longer than that, it's unlikely to get pushed out anyways.

Test Plan: Just tweaking a knob. No functional changes.

Reviewers: #stirling, jamesbartlett, jps

Reviewed By: #stirling, jps

Subscribers: yzhao

Signed-off-by: Omid Azizi <oazizi@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10659

GitOrigin-RevId: 5880e6260f5746aa87426ff2f017bf6abd884983