PerfProfiler: fix bug java_symbolizer.cc where overloaded ! operator was applied to a pointer vs. underlying.
Summary: We fix our check that tests if the symbol file was successfully opened. The ! operator was intended to apply to the object underlying the pointer, but instead was applied to the pointer. Here, we make our check more explicit by calling the fail() member function.

Test Plan: The PerfProfiler is not yet enabled in production and this will be covered by upcoming tests.

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

JIRA Issues: PP-2982

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10674

GitOrigin-RevId: 21a86369f8f56c709916935f88a322285303d3d6