[PC-1417]: Always use key field in GQL queries
Summary:
Follow-up from fixing one of these earlier this week.

When we make a GQL query, Apollo caches the result. To reconcile results of multiple queries that reference the same object, it relies on the `id` field (this can be customized in `typePolicies`; we do that already for `AutocompleteField`).

Always requesting the key field makes sure Apollo can correctly match its cache records and reduce network calls a bit. We'll need to update the fetch policy on each of them to reduce this further, however.

This diff also makes the formatting of our tagged `gql` strings consistent since I was going over all of them anyway.

Future improvement possibilities include:
- Always add the `update` option to `useMutation` queries to update caches and trigger resyncs.
- Deduplicate calls for cluster list info and put that in a context. Let it control refetches from a single point.
- Leverage Apollo's `refetchQueries` feature (and related features) to remove the need for several extra queries.
- Go over the `fetchPolicy` and `nextFetchPolicy` of every query once more to see if more network traffic can be avoided without letting cache go stale.

Test Plan:
Sit on the live view for a bit. Nothing should be noticeably different about the `graphql` calls in the network tab of the devtools. There may be slightly fewer redundant calls for cluster lists specifically.
Same goes for admin view.
The browser console in `yarn dev` should no longer complain about any fields it can't merge - if it does, we need to update a `typePolicies` entry to fix it; everything should have this covered already though.

Reviewers: michelle, vihang

Reviewed By: vihang

JIRA Issues: PC-1417

Signed-off-by: Nick Lanam <nlanam@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10685

GitOrigin-RevId: 8820e09f99b92957eb8bc07a41527a6d34a2baa0