[CLEANUP] Remove the unused stirling perf testing deployment stage
Summary:
This was added for setting up a Jenkins workflow to deploy mainline to a GKE cluster and query
PEM's resource consumption on that cluster, for tracking performance changes.

There is another pipeline running that does the same thing. So this is not needed.

Test Plan: Changes in Jenkinsfile will be picked up by the Jenkins workflow and tested automatically

Reviewers: vihang

Reviewed By: vihang

Signed-off-by: yzhao1012 <yzhao@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10684

GitOrigin-RevId: 5ffa54ef4493c2d72f5cda42855392d2230d6c6d