PerfProfiler: add test case for Java symbolization at source connector level.
Summary: This is the test case for Java symbolization. In this test, we create a perf profile source connector and run a Java test app ("fib"). We expect that, if everything works, the perf profiler will observe 2x as many fib52() symbols as fib27() symbols.

Test Plan: This is a test.

Reviewers: #stirling, oazizi

Reviewed By: #stirling, oazizi

Subscribers: oazizi

JIRA Issues: PP-2982

Signed-off-by: Pete Stevenson <jps@pixielabs.ai>

Differential Revision: https://phab.corp.pixielabs.ai/D10687

GitOrigin-RevId: 18bc1329985c86fbc4cc44384cd26d11e238493b